requirements.txt
alembic==0.8.4
aniso8601==1.1.0
argh==0.26.2
blinker==1.4
coverage==4.0.3
decorator==4.0.6
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
ipython==4.0.3
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
path.py==8.1.2
pathtools==0.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.0
ptyprocess==0.5
python-dateutil==2.5.3
python-editor==0.5
pytz==2016.6.1
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.11
traitlets==4.1.0
validate-email==1.3
watchdog==0.8.3
Werkzeug==0.11.3
